chore: refactor into substreams component
